requirements.txt
panda3d==1.10.11
requests
pyglet
pypresence
DirectGuiExtension
pyttsx3
psutil
pyscreenshot
pydatanet
distro
panda3d-character-controller
colorama
limeade
PyMsgBox
elevate